move GroupFunction* classes to o.e.sh.core.internal.items.group package
